[generic.go] Take ExtractParam from KprobeArg to build ConfigBtfArg
Searches if every user defined type with ExtractParam exists as a BTF
type and stores its corresponding offset in ConfigBtfArg.

This function does a basic split on `ExtractParam` to obtain the "path"
to the required data. Then, the array is given to `btf.FindNextBTFType` to
find the offset of each element until we reach the required data.
The output is stored in EventConfig to keep the normal behaviour.

For example, if the arg 0 is `struct linux_binprm` and ExtractParam is
set to `file.f_path.dentry.d_name.name`, the output will give an array
of all the offsets from their parents as follows :
[{ offset: 96, is_pointer: 0 }, { offset: 152, is_pointer: 1 }, ...]

Signed-off-by: Tristan d'Audibert <tristan.daudibert@orange.com>