[btf_test.go] Add unit tests for FindNextBtfType with multiple cases.
This commit adds 3 tests for FindNextBtfType algorithm.
The first is `testAssertEqualBtfPath` to assert that a specific
path has the exact same btfConfig as expected.
The second, "testAssertPathIsAccessible" tries to reach the path and
asserts that no errors are raised.
The third test "testAssertErrorOnInvalidPath" asserts that the error
messages raised if the path is incorrect.

The chosen test cases have embed union/anonymous structs.

The important thing to notice in case of adding new tests is to be
aware of btf changes on different architectures. Especially for
`testAssertEqualBtfPath` where Offset could be different.

To Test locally :
```
go test -exec "sudo" ./pkg/btf/
```

Signed-off-by: Tristan d'Audibert <tristan.daudibert@orange.com>